Initial P4PR1UM Support
Meant for reservation of the incredibly expensive, rare, cartridges, that have cost hundreds upon hundreds of USD to obtain.  No actual assets from said game are included in this code base.  It is ultimately up to the End User to legally dump their own cartridges.  More Fix-Ups to come, related to this addition.

And, official Devs of the game.  Thank you so very much for the incredible game.  Now, I can let it rest on my shelf like those action figures that are never meant to be outside of their boxes!  Thanks:)

And, thanks to ALL who have had a part in reverse engineering the protection.

Legal Precedents:

Sega v. Accolade:
.
This case established that reverse engineering for the purpose of achieving interoperability can be considered fair use, even if it involves copying some of the code.
Atari Games Corp. v. Nintendo of America:
.
This case also affirmed that reverse engineering could be permissible under certain circumstances, particularly when it is the only way to access functional elements of a program.

Again, this is primarily meant to protect this beloved gem of a game, that myself and others, will likely never be able to obtain a retail copy of, again, due to supply and demand.

Once, the digital version, if and when, appears on Next Gen, I will be the first in line to purchase!