P4PR1UM Tier Level Memory MP3 Fix
MP3s, the way they buffer and unload, affect virtually all platforms in some way, shape, or another.  This is a solid fix-up, for the interim, to help offset the memory costs, depending on RAM limitations.  I will do more commits, to further help the cause, as far as memory fixing.  Stay tuned!